# Unify fingerprinting aka signature calculation.