@beorn7 2015-03-03
Old benchmark:
BenchmarkMetric   200000              6262 ns/op            1800 B/op         32 allocs/op

New benchmark:
BenchmarkMetric   300000              5117 ns/op            1393 B/op          9 allocs/op

Note that the 9 remaining allocations are all done by the test setup. The benchmark is not particularly well written. This PR adds a better benchmark:

BenchmarkMetricLabelsToFingerprintScalar        200000000                6.34 ns/op            0 B/op          0 allocs/op
BenchmarkMetricLabelsToFingerprintSingle         5000000               354 ns/op               0 B/op          0 allocs/op
BenchmarkMetricLabelsToFingerprintDouble         3000000               454 ns/op               0 B/op          0 allocs/op
BenchmarkMetricLabelsToFingerprintTriple         3000000               562 ns/op               0 B/op          0 allocs/op